[TRANSFORMATIONS] Move convert_pagedattn_inputs to paged_attention dir (#35471)
### Details:
- moving the implementation and references of this transformation from
the `common_optimizations` directory to a new `paged_attention`
directory,
- simplify the pattern and prepare it for adding the new LA operations
(see #35128)

### Tickets:
[CVS-183796](https://jira.devtools.intel.com/browse/CVS-183796)
[CVS-183797](https://jira.devtools.intel.com/browse/CVS-183797)

### AI Assistance:
 - *AI assistance used: no*
- *If yes, summarize how AI was used and what human validation was
performed (build/tests/manual checks).*

---------
