+++ src/common/transformations/tests/common_optimizations/convert_pagedattn_inputs.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "transformations/common_optimizations/convert_pagedattn_inputs.hpp"
+#include "transformations/paged_attention/convert_pagedattn_inputs.hpp"
 
 #include <gtest/gtest.h>
